Added pipelines to new folder
