+++ scripts/tess/pipelines/.gitignore
@@ -1,0 +1,3 @@
+./tiny_qlp
+./tiny_tglc
+./tiny_spoc
