+++ scripts/tess/pipelines/tglc.py
@@ -1,19 +1,15 @@
+from base_downloader import TESS_Downloader
 import datasets
 from datasets import Features, Value, Sequence
 from datasets.data_files import DataFilesPatternsDict
 import numpy as np 
 import itertools
 import h5py
-
-# Some data cleaning is required here especially for the NaN values.
-
-# TESS Sectors
-PM = list(range(1, 26 + 1)) # Primary Mission
-EM1 = list(range(27, 55 + 1)) # Extended Mission 1
-EM2 = list(range(56, 96 + 1)) # Extended Mission 2
+from astropy.table.row import Row
+import os 
+from astropy.io import fits
 
 _CITATION = """\
-
 @ARTICLE{2023AJ....165...71H,
        author = {{Han}, Te and {Brandt}, Timothy D.},
         title = "{TESS-Gaia Light Curve: A PSF-based TESS FFI Light-curve Product}",
@@ -33,6 +29,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 """
+
 _DESCRIPTION = """
 TESS-Gaia Light Curve: Light curves from TESS FFI images using GAIA based priors.
 """
@@ -41,22 +38,21 @@ _LICENSE = "CC BY 4.0"
 _VERSION = "0.0.1"
 
 
-class TESS(datasets.GeneratorBasedBuilder):
+class TGLC(datasets.GeneratorBasedBuilder):
     VERSION = "0.0.1"
     DEFAULT_CONFIG_NAME = "all"
 
     BUILDER_CONFIGS = [
     datasets.BuilderConfig(
         name="all",
-        version=VERSION,
+        version=_VERSION,
         data_files=DataFilesPatternsDict.from_patterns(
             {"train": ["./TGLC/healpix=*/*.hdf5"]} 
         ),
-        description="TGLC light curves (S0023)",
+        description="TGLC light curves",
         )
     ]
 
-    
     @classmethod
     def _info(self):
         features = {"lightcurve" : Sequence({
@@ -70,7 +66,7 @@ class TESS(datasets.GeneratorBasedBuilder):
             'RA':  Value(dtype="float32"),
             'DEC':  Value(dtype="float32"),
             'TIC_ID': Value(dtype="string"),
-            'gaiadr3_id': Value(dtype="string"),
+            'GAIADR3_ID': Value(dtype="string"),
             'aper_flux_err':  Value(dtype="float32"),
             'psf_flux_err': Value(dtype="float32"),
         }
@@ -139,8 +135,120 @@ class TESS(datasets.GeneratorBasedBuilder):
                         'RA':  data["RA"][i],
                         'DEC':  data["DEC"][i],
                         'TIC_ID': data["TIC_ID"][i],
-                        'gaiadr3_id': data["gaiadr3_id"][i],
+                        'GAIADR3_ID': data["GAIADR3_ID"][i],
                         'psf_flux_err': data["psf_flux_err"][i],
                         'aper_flux_err':  data["aper_flux_err"][i]
                     }
                     yield str(data["TIC_ID"][i]), example
+
+class TGLC_Downloader(TESS_Downloader):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        if self.sector > 39:
+            raise ValueError(f"TGLC is currently only available for sectors 1-39.") 
+
+        self._pipeline = 'TGLC'
+        self.fits_base_url =  f'https://archive.stsci.edu/hlsps/tglc/{self.sector_str}/'         
+        self._sh_base_url = f'https://archive.stsci.edu/hlsps/tglc/download_scripts/hlsp_tglc_tess_ffi_{self.sector_str}_tess_v1_llc.sh'
+        self._csv_base_url = f'https://archive.stsci.edu/hlsps/tglc/target_lists/{self.sector_str}.csv'
+        self._catalog_column_names = ['GAIADR3_ID', 'cam', 'ccd', 'fp1', 'fp2', 'fp3', 'fp4'] # put the joining id first!
+        
+    @property
+    def pipeline(self): # TESS-Pipeline
+        return self._pipeline
+    
+    @property
+    def csv_url(self): 
+        return self._csv_base_url
+    
+    @property
+    def sh_url(self):
+        return self._sh_base_url
+
+    @property
+    def catalog_column_names(self):
+        return self._catalog_column_names
+    
+    # Add methods - processing_fn, fits_url, parse_line
+    def parse_line(self, line: str) -> list[int]:
+        '''
+        Parse a line from the .sh file, extract the relevant parts (gaia_id, cam, ccd, fp1, fp2, fp3, fp4) and return them as a list
+
+        Parameters
+        ----------
+        line: str, a line from the .sh file
+        
+        Returns
+        ------- 
+        params: list, list of parameters extracted from the line
+        '''
+
+        parts = line.split()
+        output_path = parts[4].strip("'")
+            
+        # Split the path and extract the relevant parts
+        path_parts = output_path.split('/')
+        
+        cam = int(path_parts[1].split('-')[0].split('cam')[1])
+        ccd = int(path_parts[1].split('-')[1].split('ccd')[1])
+        numbers = path_parts[2:6]
+        gaia_id = path_parts[-1].split('-')[1]
+
+        return [int(gaia_id), cam, ccd, *numbers]
+    
+    def fits_url(self, catalog_row: Row) -> tuple[str, str]:
+        path = f'cam{catalog_row["cam"]}-ccd{catalog_row["ccd"]}/{catalog_row["fp1"]}/{catalog_row["fp2"]}/{catalog_row["fp3"]}/{catalog_row["fp4"]}/hlsp_tglc_tess_ffi_gaiaid-{catalog_row["GAIADR3_ID"]}-{self.sector_str}-cam{catalog_row["cam"]}-ccd{catalog_row["ccd"]}_tess_v1_llc.fits'
+        url =  self.fits_base_url + path
+        return url, path
+    
+    def processing_fn(
+            self,
+            catalog_row : Row
+    ) -> dict:
+        ''' 
+        Process a single light curve file into the standard format 
+
+        Parameters
+        ----------
+        row: astropy Row, a single row from the sector catalog containing the object descriptors: gaiadr3_id, cam, ccd, fp1, fp2, fp3, fp4
+        
+        Returns
+        ------- 
+        lightcurve: dict, light curve in the standard format
+        i.e. 
+            {
+                'TIC_ID': tess id,
+                'gaiadr3_id': gaia id,
+                'time': obs_times: arr_like,
+                'psf_flux': psf_fluxes: arr_like,
+                'psf_flux_err': psf_flux_err: float,
+                'aper_flux': aperture_fluxes: arr_like,
+                'aper_flux_err': aperture_flux_err: float,
+                'tess_flags': tess_flags: arr_like,
+                'tglc_flags': tglc_flags: arr_like,
+                'RA': ra: float,
+                'DEC': dec # More columns maybe required...
+            }
+        '''
+        fits_fp = os.path.join(self.fits_dir, self.fits_url(catalog_row)[1])
+
+        try:
+            with fits.open(fits_fp, mode='readonly', memmap=True) as hdu:
+        
+                return {'TIC_ID': catalog_row['TIC_ID'],
+                        'GAIADR3_ID': catalog_row['GAIADR3_ID'],
+                        'time': hdu[1].data['time'],
+                        'psf_flux': hdu[1].data['psf_flux'],
+                        'psf_flux_err': hdu[1].header['psf_err'],
+                        'aper_flux': hdu[1].data['aperture_flux'],
+                        'aper_flux_err': hdu[1].header['aper_err'],
+                        'tess_flags': hdu[1].data['TESS_flags'],
+                        'tglc_flags': hdu[1].data['TGLC_flags'],
+                        'RA': hdu[1].header['ra_obj'],
+                        'DEC': hdu[1].header['dec_obj']
+                        }
+            
+        except FileNotFoundError:
+            print(f"File not found: {fits_fp}")
+            return None
