--- scripts/tess/qlp/.gitignore
@@ -1,1 +1,0 @@
-./qlp
