--- scripts/tess/tglc/.gitignore
@@ -1,1 +1,0 @@
-./tglc_data
