--- scripts/tess/tglc/requirements.txt
@@ -1,83 +1,0 @@
-aiohappyeyeballs==2.4.3
-aiohttp==3.10.10
-aiosignal==1.3.1
-appnope==0.1.4
-astropy==6.1.4
-astropy-iers-data==0.2024.10.21.0.33.21
-astroquery==0.4.7
-asttokens==2.4.1
-attrs==24.2.0
-beautifulsoup4==4.12.3
-certifi==2024.8.30
-charset-normalizer==3.4.0
-comm==0.2.2
-contourpy==1.3.0
-cycler==0.12.1
-datasets==3.0.2
-debugpy==1.8.7
-decorator==5.1.1
-dill==0.3.8
-executing==2.1.0
-filelock==3.16.1
-fonttools==4.54.1
-frozenlist==1.5.0
-fsspec==2024.9.0
-h5py==3.12.1
-healpy==1.17.3
-html5lib==1.1
-huggingface-hub==0.26.1
-idna==3.10
-iniconfig==2.0.0
-ipykernel==6.29.5
-ipython==8.29.0
-jaraco.classes==3.4.0
-jaraco.context==6.0.1
-jaraco.functools==4.1.0
-jedi==0.19.1
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-keyring==25.4.1
-kiwisolver==1.4.7
-matplotlib==3.9.2
-matplotlib-inline==0.1.7
-more-itertools==10.5.0
-multidict==6.1.0
-multiprocess==0.70.16
-nest-asyncio==1.6.0
-numpy==2.1.2
-packaging==24.1
-pandas==2.2.3
-parso==0.8.4
-pexpect==4.9.0
-pillow==11.0.0
-platformdirs==4.3.6
-pluggy==1.5.0
-prompt_toolkit==3.0.48
-propcache==0.2.0
-psutil==6.1.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pyarrow==17.0.0
-pyerfa==2.0.1.4
-Pygments==2.18.0
-pyparsing==3.2.0
-pytest==8.3.3
-python-dateutil==2.9.0.post0
-pytz==2024.2
-pyvo==1.5.3
-PyYAML==6.0.2
-pyzmq==26.2.0
-requests==2.32.3
-six==1.16.0
-soupsieve==2.6
-stack-data==0.6.3
-tornado==6.4.1
-tqdm==4.66.5
-traitlets==5.14.3
-typing_extensions==4.12.2
-tzdata==2024.2
-urllib3==2.2.3
-wcwidth==0.2.13
-webencodings==0.5.1
-xxhash==3.5.0
-yarl==1.16.0
